feat: Add all visible NPC to save game.
