+++ Assets/UnZENity-Core/Scripts/Extensions/UnityExtension.cs
@@ -1,4 +1,5 @@
 using UnityEngine;
+using ZenKit.Util;
 
 namespace GUZ.Core.Extensions
 {
@@ -21,6 +22,25 @@ namespace GUZ.Core.Extensions
                 return vector * 100;
             else
                 return vector;
+        }
+        
+        // Create back conversion from UnityQuaternion to Matrix3x3
+        public static Matrix3x3 ToZkMatrix(this Quaternion quaternion)
+        {
+            var unityMatrix = Matrix4x4.Rotate(quaternion);
+
+            return new Matrix3x3(
+                m11: unityMatrix.m00,
+                m12: unityMatrix.m01,
+                m13: unityMatrix.m02,
+
+                m21: unityMatrix.m10,
+                m22: unityMatrix.m11,
+                m23: unityMatrix.m12,
+
+                m31: unityMatrix.m20,
+                m32: unityMatrix.m21,
+                m33: unityMatrix.m22);
         }
     }
 }
