+++ Assets/UnZENity-Core/Scripts/Npc/NpcLoader.cs
@@ -1,3 +1,5 @@
+using GUZ.Core.Data.Container;
+using GUZ.Core.Extensions;
 using UnityEngine;
 using ZenKit.Daedalus;
 
@@ -6,6 +8,7 @@ namespace GUZ.Core.Npc
     public class NpcLoader : MonoBehaviour
     {
         public NpcInstance Npc;
+        public NpcContainer Container => Npc.GetUserData();
         public bool IsLoaded;
     }
 }
